requirements.txt
Django
djangorestframework
requests
ratelimit
Celery
